backend/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
icecream==2.1.3
pillow==10.2.0
Pygments==2.17.2
PyJWT==2.8.0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
